+++ VTK/Graphics/vtkStreamPoints.cxx
@@ -42,7 +42,7 @@ void vtkStreamPoints::Execute()
   vtkCellArray *newVerts;
   vtkIdType i, ptId, id;
   int j;
-  float tOffset, x[3], v[3], s, r;
+  double tOffset, x[3], v[3], s, r;
   vtkPolyData *output = this->GetOutput();
   vtkDataSet *input = this->GetInput();
   vtkIdList *pts;
